feat: extracted to public pkg templateprovider
