+++ internal/filetreeprocessor/filetreeprocessor.go
@@ -2,7 +2,7 @@ package filetreeprocessor
 
 import (
 	"github.com/apex/log"
-	"github.com/gchiesa/ska/internal/templateprovider"
+	"github.com/gchiesa/ska/pkg/templateprovider"
 	"os"
 )
