+++ internal/filetreeprocessor/type.go
@@ -3,7 +3,7 @@ package filetreeprocessor
 import (
 	"github.com/apex/log"
 	"github.com/gchiesa/ska/internal/multipart"
-	"github.com/gchiesa/ska/internal/templateprovider"
+	"github.com/gchiesa/ska/pkg/templateprovider"
 )
 
 type FileTreeProcessor struct {
